# perf(gzip): switch to flate2