@domodwyer 2020-11-30
I can follow this up with another PR(s) to switch other bits of the codebase using `libflate` if we're happy.