@domodwyer 2020-11-30
@alamb experimentally it doesn't seem to be required - I haven't checked, but I expect it performs internal buffering. There's no performance change when adding a `BufReader` to a `flate2` decoder so I wouldn't bother.